Add support for custom filename calculator in rotating_file_sink.
