+++ tests/CMakeLists.txt
@@ -18,6 +18,7 @@ set(SPDLOG_UTESTS_SOURCES
     test_file_helper.cpp
     test_file_logging.cpp
     test_daily_logger.cpp
+    test_rotating_logger.cpp
     test_misc.cpp
     test_eventlog.cpp
     test_pattern_formatter.cpp
